+++ packages/features/ee/workflows/api/scheduleEmailReminders.ts
@@ -1,5 +1,4 @@
 /* Schedule any workflow reminder that falls within 72 hours for email */
-import type { Prisma, WorkflowReminder, WorkflowStep } from "@prisma/client";
 import client from "@sendgrid/client";
 import sgMail from "@sendgrid/mail";
 import type { NextApiRequest, NextApiResponse } from "next";
@@ -11,10 +10,15 @@ import logger from "@calcom/lib/logger";
 import { defaultHandler } from "@calcom/lib/server";
 import { getTimeFormatStringFromUserTimeFormat } from "@calcom/lib/timeFormat";
 import prisma from "@calcom/prisma";
-import type { EventType, User } from "@calcom/prisma/client";
 import { WorkflowActions, WorkflowMethods, WorkflowTemplates } from "@calcom/prisma/enums";
 import { bookingMetadataSchema } from "@calcom/prisma/zod-utils";
 
+import type { PartialWorkflowReminder } from "../lib/getWorkflowReminders";
+import {
+  getAllRemindersToCancel,
+  getAllRemindersToDelete,
+  getAllUnscheduledReminders,
+} from "../lib/getWorkflowReminders";
 import { getiCalEventAsString } from "../lib/getiCalEventAsString";
 import type { VariablesType } from "../lib/reminders/templates/customTemplate";
 import customTemplate from "../lib/reminders/templates/customTemplate";
@@ -25,33 +29,6 @@ const senderEmail = process.env.SENDGRID_EMAIL as string;
 
 sgMail.setApiKey(sendgridAPIKey);
 client.setApiKey(sendgridAPIKey);
-
-type PartialWorkflowStep = Partial<WorkflowStep> | null;
-
-type Booking = Prisma.BookingGetPayload<{
-  include: {
-    attendees: true;
-  };
-}>;
-
-type PartialBooking =
-  | (Pick<
-      Booking,
-      | "startTime"
-      | "endTime"
-      | "location"
-      | "description"
-      | "metadata"
-      | "customInputs"
-      | "responses"
-      | "uid"
-      | "attendees"
-    > & { eventType: Partial<EventType> | null } & { user: Partial<User> | null })
-  | null;
-
-type PartialWorkflowReminder = Pick<WorkflowReminder, "id" | "scheduledDate"> & {
-  booking: PartialBooking | null;
-} & { workflowStep: PartialWorkflowStep };
 
 async function handler(req: NextApiRequest, res: NextApiResponse) {
   const apiKey = req.headers.authorization || req.query.apiKey;
@@ -68,35 +45,7 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
   const sandboxMode = process.env.NEXT_PUBLIC_IS_E2E ? true : false;
 
   // delete batch_ids with already past scheduled date from scheduled_sends
-  const pageSize = 90;
-  let pageNumber = 0;
-
-  const remindersToDelete: { referenceId: string | null }[] = [];
-
-  while (true) {
-    const newRemindersToDelete = await prisma.workflowReminder.findMany({
-      where: {
-        method: WorkflowMethods.EMAIL,
-        cancelled: true,
-        scheduledDate: {
-          lte: dayjs().toISOString(),
-        },
-      },
-      skip: pageNumber * pageSize,
-      take: pageSize,
-      select: {
-        referenceId: true,
-      },
-    });
-
-    if (newRemindersToDelete.length === 0) {
-      break;
-    }
-
-    remindersToDelete.push(...newRemindersToDelete);
-
-    pageNumber++;
-  }
+  const remindersToDelete: { referenceId: string | null }[] = await getAllRemindersToDelete();
 
   const deletePromises: Promise<any>[] = [];
 
@@ -127,35 +76,7 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
   });
 
   //cancel reminders for cancelled/rescheduled bookings that are scheduled within the next hour
-  pageNumber = 0;
-
-  const remindersToCancel: { referenceId: string | null; id: number }[] = [];
-
-  while (true) {
-    const newRemindersToCancel = await prisma.workflowReminder.findMany({
-      where: {
-        cancelled: true,
-        scheduled: true, //if it is false then they are already cancelled
-        scheduledDate: {
-          lte: dayjs().add(1, "hour").toISOString(),
-        },
-      },
-      skip: pageNumber * pageSize,
-      take: pageSize,
-      select: {
-        referenceId: true,
-        id: true,
-      },
-    });
-
-    if (newRemindersToCancel.length === 0) {
-      break;
-    }
-
-    remindersToCancel.push(...newRemindersToCancel);
-
-    pageNumber++;
-  }
+  const remindersToCancel: { referenceId: string | null; id: number }[] = await getAllRemindersToCancel();
 
   const cancelUpdatePromises: Promise<any>[] = [];
 
@@ -190,84 +111,13 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
   });
 
   // schedule all unscheduled reminders within the next 72 hours
-  pageNumber = 0;
   const sendEmailPromises: Promise<any>[] = [];
 
-  const unscheduledReminders: PartialWorkflowReminder[] = [];
+  const unscheduledReminders: PartialWorkflowReminder[] = await getAllUnscheduledReminders();
 
-  while (true) {
-    const newUnscheduledReminders = await prisma.workflowReminder.findMany({
-      where: {
-        method: WorkflowMethods.EMAIL,
-        scheduled: false,
-        scheduledDate: {
-          lte: dayjs().add(72, "hour").toISOString(),
-        },
-        OR: [{ cancelled: false }, { cancelled: null }],
-      },
-      skip: pageNumber * pageSize,
-      take: pageSize,
-      select: {
-        id: true,
-        scheduledDate: true,
-        workflowStep: {
-          select: {
-            action: true,
-            sendTo: true,
-            reminderBody: true,
-            emailSubject: true,
-            template: true,
-            sender: true,
-            includeCalendarEvent: true,
-          },
-        },
-        booking: {
-          select: {
-            startTime: true,
-            endTime: true,
-            location: true,
-            description: true,
-            user: {
-              select: {
-                email: true,
-                name: true,
-                timeZone: true,
-                locale: true,
-                username: true,
-                timeFormat: true,
-                hideBranding: true,
-              },
-            },
-            metadata: true,
-            uid: true,
-            customInputs: true,
-            responses: true,
-            attendees: true,
-            eventType: {
-              select: {
-                bookingFields: true,
-                title: true,
-                slug: true,
-                recurringEvent: true,
-              },
-            },
-          },
-        },
-      },
-    });
-
-    if (!newUnscheduledReminders.length && pageNumber === 0) {
-      res.status(200).json({ message: "No Emails to schedule" });
-      return;
-    }
-
-    if (newUnscheduledReminders.length === 0) {
-      break;
-    }
-
-    unscheduledReminders.push(...newUnscheduledReminders);
-
-    pageNumber++;
+  if (!unscheduledReminders.length) {
+    res.status(200).json({ message: "No Emails to schedule" });
+    return;
   }
 
   for (const reminder of unscheduledReminders) {
@@ -385,7 +235,7 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
         const batchId = batchIdResponse[1].batch_id;
 
         if (reminder.workflowStep.action !== WorkflowActions.EMAIL_ADDRESS) {
-          await sendEmailPromises.push(
+          sendEmailPromises.push(
             sgMail.send({
               to: sendTo,
               from: {
@@ -431,6 +281,14 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
       logger.error(`Error scheduling Email with error ${error}`);
     }
   }
+
+  Promise.allSettled(sendEmailPromises).then((results) => {
+    results.forEach((result) => {
+      if (result.status === "rejected") {
+        console.log("Email sending failed", result.reason);
+      }
+    });
+  });
 
   res.status(200).json({ message: `${unscheduledReminders.length} Emails scheduled` });
 }
